feat: OAuthAuthenticator 클래스 구현
